requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
pymongo==4.5.0
PyJWT==2.8.0
python-dotenv==1.0.0